Set token handler on ms2
